# docs/swagger.yaml
basePath: /
definitions:
...
  customerAuth.Address:
    properties:
      add:
        type: string
      coordinates:
        items:
          type: number
        type: array
      type:
        type: string
    type: object
...
  customerAuth.CustomerResDto:
    properties:
      address:
        $ref: '#/definitions/customerAuth.Address'
      age:
        type: string
      createdAt:
        type: string
      email:
        type: string
      familyMembers:
        items:
          $ref: '#/definitions/customerAuth.FamilyMembers'
        type: array
      firstName:
        type: string
      id:
        type: string
      image:
        type: string
      lastName:
        type: string
      notification:
        $ref: '#/definitions/customerAuth.Notification'
      phoneNumber:
        $ref: '#/definitions/customerAuth.PhoneNumber'
      sex:
        type: string
      updatedAt:
        type: string
    type: object
...
  customerAuth.FamilyMembers:
    properties:
      age:
        type: string
      id:
        type: string
      name:
        type: string
      relationShip:
        type: string
      sex:
        type: string
    type: object
  customerAuth.GetCustomerResDto:
    properties:
      data:
        $ref: '#/definitions/customerAuth.CustomerResDto'
      message:
        type: string
      status:
        type: boolean
    type: object
...
  customerAuth.Notification:
    properties:
      deviceToken:
        type: string
      deviceType:
        type: string
      isEnabled:
        type: boolean
    type: object
  customerAuth.PhoneNumber:
    properties:
      countryCode:
        type: string
      dialCode:
        type: string
      number:
        type: string
    type: object
...
  services.DoctorRes:
    properties:
      id:
        type: string
      image:
        type: string
      name:
        type: string
      schedule:
        items:
          $ref: '#/definitions/services.DoctorScheduleRes'
        type: array
      speciality:
        type: string
    type: object
  services.DoctorResDto:
    properties:
      data:
        items:
          $ref: '#/definitions/services.SpecialityDoctorsRes'
        type: array
      message:
        type: string
      status:
        type: boolean
    type: object
  services.DoctorScheduleRes:
    properties:
      days:
        items:
          type: string
        type: array
      endTime:
        type: string
      startTime:
        type: string
    type: object
...
  services.MedicalLabScientistResDto:
    properties:
      data:
        $ref: '#/definitions/services.Role'
      message:
        type: string
      status:
        type: boolean
    type: object
...
  services.Role:
    properties:
      facilityOrProfession:
        type: string
      image:
        type: string
      isEmergencyAvailable:
        type: boolean
      name:
        type: string
      role:
        type: string
      serviceStatus:
        type: string
    type: object
...
  services.SpecialityDoctorsRes:
    properties:
      doctors:
        items:
          $ref: '#/definitions/services.DoctorRes'
        type: array
      speciality:
        type: string
    type: object
  services.SpecialityTrainerRes:
    properties:
      category:
        type: string
      trainers:
        items:
          $ref: '#/definitions/services.TrainerRes'
        type: array
    type: object
...
  services.TrainerRes:
    properties:
      category:
        type: string
      id:
        type: string
      information:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  services.TrainerResDto:
    properties:
      data:
        items:
          $ref: '#/definitions/services.SpecialityTrainerRes'
        type: array
      message:
        type: string
      status:
        type: boolean
    type: object
...
host: mmpn2atpcm.eu-west-1.awsapprunner.com
info:
  contact:
    name: Chirag Sharma
  description: Care Ville Backend (in GoLang)
  license:
    name: MIT
  title: Care Ville
  version: 0.0.1
paths:
...
  /customer/get-customer-info:
    get:
      consumes:
      - application/json
      description: Get customer by ID
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customerAuth.GetCustomerResDto'
      summary: Get customer by ID
      tags:
      - customer authorization
...
  /provider/services/add-medicalLab-scientist:
    post:
      consumes:
      - multipart/form-data
      description: Add MedicalLabScientist
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      - description: physiotherapistImage
        in: formData
        name: medicalLabScientistImage
        type: file
      - description: professionalCertificate
        in: formData
        name: professionalCertificate
        type: file
      - description: professionalLicense
        in: formData
        name: professionalLicense
        type: file
      - description: personalLicense
        in: formData
        name: personalLicense
        type: file
      - description: personalNimc
        in: formData
        name: personalNimc
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.MedicalLabScientistResDto'
      summary: Add MedicalLabScientist
      tags:
      - medicalLabScientist
...
  /provider/services/get-all-doctors:
    get:
      consumes:
      - application/json
      description: GetAllDoctors
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.DoctorResDto'
      summary: GetAllDoctors
      tags:
      - hospClinic
...
  /provider/services/get-all-trainers:
    get:
      consumes:
      - application/json
      description: Get all trainers
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TrainerResDto'
      summary: Get all trainers
      tags:
      - fitnessCenter
...
swagger: "2.0"